+++ gpslocation.py
@@ -49,14 +49,33 @@ if (gps_enabled == True and config["general"]["gps"]["provider"] == "gpsd"): # O
 
 
 
+def calculate_speed(location_history, hardware_speed):
+    speed = 0 # This is a placeholder that will be replaced with the actual speed.
+    if (config["general"]["gps"]["speed_source"] == "calculated"): # The configuration indicates that the speed should be calculated based on distance and time.
+        if (len(location_history) >= 2):
+            distance = get_distance(location_history[-1]["lat"], location_history[-1]["lon"], location_history[-2]["lat"], location_history[-2]["lon"]) # Get the distance between the two points.
+            time_difference = abs(location_history[-1]["time"] - location_history[-2]["time"]) # Get the time difference between the two points.
+            miles_per_second = distance / time_difference
+            speed = miles_per_second * 1609.344
+        else:
+            speed = 0
+    elif (config["general"]["gps"]["speed_source"] == "hardware"):
+        speed = hardware_speed
+    else: # This will only happen when the configuration is invalid.
+        speed = hardware_speed
+
+    return speed
+
+
 # Define the function that will be used to get the current GPS coordinates.
 debug_message("Creating `get_gps_location` function")
-def get_gps_location():
+def get_gps_location(location_history=[]):
     debug_message("Getting GPS location")
     if (gps_enabled == True): # Check to see if GPS is enabled.
         if (config["general"]["gps"]["demo_mode"]["enabled"] == True): # Check to see if GPS demo mode is enabled in the configuration.
             debug_message("Returning demo GPS information")
-            return float(config["general"]["gps"]["demo_mode"]["data"]["longitude"]), float(config["general"]["gps"]["demo_mode"]["data"]["latitude"]), float(config["general"]["gps"]["demo_mode"]["data"]["speed"]), float(config["general"]["gps"]["demo_mode"]["data"]["altitude"]), float(config["general"]["gps"]["demo_mode"]["data"]["heading"]), int(config["general"]["gps"]["demo_mode"]["data"]["satellites"]), "V0LT Assassin - GPS demo mode" # Return the sample GPS information defined in the configuration.
+            speed = calculate_speed(location_history, config["general"]["gps"]["demo_mode"]["data"]["speed"])
+            return float(config["general"]["gps"]["demo_mode"]["data"]["longitude"]), float(config["general"]["gps"]["demo_mode"]["data"]["latitude"]), speed, float(config["general"]["gps"]["demo_mode"]["data"]["altitude"]), float(config["general"]["gps"]["demo_mode"]["data"]["heading"]), int(config["general"]["gps"]["demo_mode"]["data"]["satellites"]), "V0LT Assassin - GPS demo mode" # Return the sample GPS information defined in the configuration.
         else: # GPS demo mode is disabled, so attempt to get the actual GPS data from GPSD.
             if (config["general"]["gps"]["provider"] == "gpsd"):
                 try: # Don't terminate the entire script if the GPS location fails to be aquired.
@@ -65,7 +84,8 @@ def get_gps_location():
                     debug_message("Fetching GPSD information")
                     gps_data_packet = gpsd.get_current() # Get the current information.
                     debug_message("Received GPSD information")
-                    return gps_data_packet.position()[0], gps_data_packet.position()[1], gps_data_packet.speed(), gps_data_packet.altitude(), gps_data_packet.movement()["track"], gps_data_packet.sats, "V0LT Assassin - GPSD Device" # Return GPS information.
+                    speed = calculate_speed(location_history, gps_data_packet.speed())
+                    return gps_data_packet.position()[0], gps_data_packet.position()[1], speed, gps_data_packet.altitude(), gps_data_packet.movement()["track"], gps_data_packet.sats, "V0LT Assassin - GPSD Device" # Return GPS information.
                 except: # If the current location can't be established, then return placeholder location data.
                     return [0.0000, -0.0000, 0.0, 0.0, 0.0, 0, "V0LT Assassin"] # Return a default placeholder location.
                     debug_message("GPS fetch failed")
@@ -75,7 +95,8 @@ def get_gps_location():
                     raw_termux_response = str(os.popen("termux-location").read()) # Execute the Termux location command.
                     termux_response = json.loads(raw_termux_response) # Load the location information from the Termux response.
                     debug_message("Received termux-location information")
-                    return termux_response["latitude"], termux_response["longitude"], termux_response["speed"], termux_response["altitude"], termux_response["bearing"], 0, "V0LT Assassin - Termux-API" # Return the fetched GPS information.
+                    speed = calculate_speed(location_history, termux_response["speed"])
+                    return termux_response["latitude"], termux_response["longitude"], speed, termux_response["altitude"], termux_response["bearing"], 0, "V0LT Assassin - Termux-API" # Return the fetched GPS information.
                 except:
                     return [0.0000, -0.0000, 0.0, 0.0, 0.0, 0, "V0LT Assassin"] # Return a default placeholder location.
                     debug_message("GPS fetch failed")
@@ -86,7 +107,8 @@ def get_gps_location():
                     locateme_response = raw_locateme_response.split(",") # Load the location information from the LocateMe response.
                     locateme_respose.append("V0LT Assassin - MacOS Location Services via LocateMe") # Append the location provider to the information.
                     debug_message("Received LocateMe information")
-                    return float(locateme_response[0]), float(locateme_response[1]), float(locateme_response[2]), float(locateme_response[3]), float(locateme_response[4]), 0 # Return the fetched GPS information.
+                    speed = calculate_speed(location_history, locateme_response[2])
+                    return float(locateme_response[0]), float(locateme_response[1]), speed, float(locateme_response[3]), float(locateme_response[4]), 0 # Return the fetched GPS information.
                 except:
                     return [0.0000, -0.0000, 0.0, 0.0, 0.0, 0, "V0LT Assassin"] # Return a default placeholder location.
                     debug_message("GPS fetch failed")